+++ flake.nix
@@ -3,7 +3,7 @@
 
   outputs = { self, nixpkgs }:
     let
-      package = "hmenu";
+      package = "hdmenu";
       system  = "x86_64-linux";
       pkgs    = nixpkgs.legacyPackages.${system};
       overlay = final: prev: {